+++ src/main/java/ksyun/client/bill_union/describesplititemdaybilldetails.v20200101/DescribeSplitItemDayBillDetailsRequest.java
@@ -34,7 +34,7 @@ public class DescribeSplitItemDayBillDetailsRequest{
     /**主账号所托管的账号Id
 */
     @KsYunField(name="SubAccountId")
-    private Integer SubAccountId;
+    private Long SubAccountId;
 
     /**第几页*/
     @KsYunField(name="Page")
