+++ src/main/java/ksyun/client/epc/createremotemanagement.v20151101/CreateRemoteManagementRequest.java
@@ -32,6 +32,6 @@ public class CreateRemoteManagementRequest{
 
     /**版本ID*/
     @KsYunField(name="VersionId")
-    private Integer VersionId;
+    private Long VersionId;
 
 }
