+++ src/main/java/ksyun/client/epc/describeinspections.v20151101/DescribeInspectionsResponse.java
@@ -25,7 +25,7 @@ public class DescribeInspectionsResponse extends BaseResponseModel {
 
     /**裸金属服务器总数*/
     @JsonProperty("TotalCount")
-    private Integer TotalCount;
+    private Long TotalCount;
 
     /**故障的信息*/
     @JsonProperty("InspectionSet")
